scripts for creating a root filesystem and ramdisk
Use these scripts directly or as an inspiration for creating a root
filesystem for backing up a QNAP drive.

Use create-sysroot.sh <SYSROOT> <QTS> to create a system root
filsystem in <SYSROOT>.

The <QTS> argument is the full path to the QNAP QTS GPL sources with
the kernel and lvm2 and thin provisioning tools. Other tools are not
required:

     ${QTS}/GPL_QTS/src/
         linux-5.10
         lvm2-2_02_138
         thin-provisioning-tools-v0.4.1

Note that lvm2 and thin-provisioning-tools are already built when
extracting the sources from the QNAP SourceForge page.

The kernel must be patched and built including modules.
Use modules_install with the INSTALL_MOD_PATH=modules option
to install the modules locally in the linux-5.10/modules directory.

Signed-off-by: Chris Zankel <chris@zankel.net>